requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
influxdb-client==1.49.0
python-dotenv==1.0.1
streamlit==1.44.1
plotly==6.0.1
pandas==2.3.3
numpy==2.2.4
pydantic==2.11.2
